Split up the code
