+++ src/Hooks/useList.js
@@ -1,32 +1,12 @@
-import { useEffect, useState } from "react"
+import useFetch from "./useFetch"
 import Api from "../Lib/Api"
 
-export default function useList(query) {
-  const [items, setItems] = useState([])
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState()
+export default function useItem(query) {
+  const result = useFetch(`${Api}/item/list/${query}`)
 
-  useEffect(
-    function () {
-      setIsLoading(true)
-      setError()
-      async function getItems() {
-        try {
-          const response = await fetch(`${Api}/item/list${query}`)
-          const items = (await response.json()).items
-          setItems(items)
-        } catch (error) {
-          console.log(error)
-          setError(error)
-        } finally {
-          setIsLoading(false)
-        }
-      }
-
-      getItems()
-    },
-    [query]
-  )
-
-  return { items, isLoading, error }
+  return {
+    items: result.json ? result.json.items : [],
+    isLoading: result.isLoading,
+    error: result.error,
+  }
 }
